requirements/test.txt
-r base.txt

coverage==3.6
django-discover-runner==0.3
